module-1/image_captioning/image_captioning/image_cap.py
# import requests
from PIL import Image
from transformers import AutoProcessor, BlipForConditionalGeneration

# ToDo: Learn AI and ML
# Load the pretrained processor and model
processor = AutoProcessor.from_pretrained("Salesforce/blip-image-captioning-base")
model = BlipForConditionalGeneration.from_pretrained("Salesforce/blip-image-captioning-base")

# Load your image, DONT FORGET TO WRITE YOUR IMAGE NAME
img_path = "/home/yguezide/projects/courses/Building Generative AI-Powered Applications with Python/image_captioning/assets/Screenshot 2023-12-25 142152.png"
# convert it into an RGB format 
image = Image.open(img_path).convert('RGB')

# You do not need a question for image captioning
text = ""
inputs = processor(images=image, text=text, return_tensors="pt")

# Generate a caption for the image
outputs = model.generate(**inputs, max_length=50)

# Decode the generated tokens to text
caption = processor.decode(outputs[0], skip_special_tokens=True)
# Print the caption
print(caption)
