module-1/image_captioning/pyproject.toml
[tool.poetry]
name = "image-captioning"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "0.1.11"
gradio = "4.44.0"
transformers = "4.38.2"
bs4 = "0.0.2"
requests = "2.31.0"
torch = "2.2.1"
pillow = "10.4"
beautifulsoup4 = "^4.12.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
